+++ packages/yourloops/test/integration/pages/signup-stepper/signup.hcp.spec.tsx
@@ -99,6 +99,12 @@ describe('Signup stepper as hcp', () => {
     fireEvent.mouseDown(within(hcpProfessionSelector).getByRole('combobox', { hidden: true }))
     await userEvent.click(screen.getByText('Nurse'))
 
+    fireEvent.mouseDown(within(screen.getByTestId('country-selector')).getByRole('combobox'))
+    fireEvent.click(screen.getByRole('option', { name: 'Japan' }))
+
+    fireEvent.mouseDown(within(screen.getByTestId('country-selector')).getByRole('combobox'))
+    fireEvent.click(screen.getByRole('option', { name: 'United Kingdom' }))
+
     expect(createButton).not.toBeDisabled()
     await userEvent.click(screen.getByText('Create account'))
 
@@ -108,7 +114,7 @@ describe('Signup stepper as hcp', () => {
         role: UserRole.Hcp,
         profile: expectedProfile,
         preferences: { displayLanguageCode: 'en' },
-        settings: { country: CountryCodes.France }
+        settings: { country: CountryCodes.UnitedKingdom }
       })
     )
     expect(getAccessTokenWithPopupMock).toHaveBeenCalledWith({ authorizationParams: { ignoreCache: true } })