+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
+## 3.3.0-terumo - 2024-06-12
+### Changed
+- YLP-3075 As a user, I want to be able to choose Japan or UK country
+- YLP-3087 As a user, I want to see the medisafe pump event in yourloops
+
+### Engineering use
+- YLP-2948 Updated failing tests because of the time change
+
+### Fixed
+- YLP-3104 Fix parameters locales issues (move the locale files in the repo)
+
 ## 3.3.0 - 2024-04-05
 ### Changed
 - YLP-2824 Display pump logos based on manufacturer